requirements.txt
pandas==2.2.0
python-dotenv==1.0.0
chardet==5.2.0
requests==2.31.0
pytest==7.4.0
python-dateutil==2.8.2
